Move duplicated code into IDynamic for #93
