+++ GearBox.Core/Model/GameObjects/ChangeTracking/ChangeTracker.cs
@@ -1,13 +1,18 @@
+using GearBox.Core.Model.Json;
+
 namespace GearBox.Core.Model.GameObjects.ChangeTracking;
 
 public class ChangeTracker
 {
     private readonly IDynamic _target;
+    private readonly Serializer _serializer;
+    private bool _hasChangedSinceLastTick = true;
     private int _lastHash;
 
     public ChangeTracker(IDynamic target)
     {
         _target = target;
+        _serializer = new Serializer("", _target.Serialize);
         _lastHash = CurrentHash();
     }
 
@@ -15,6 +20,7 @@ public class ChangeTracker
 
     public void Update()
     {
+        _hasChangedSinceLastTick = HasChanged;
         _lastHash = CurrentHash();
     }
 
@@ -27,5 +33,13 @@ public class ChangeTracker
         }
         var hashCode = result.ToHashCode();
         return hashCode;
+    }
+
+    public StableJson ToJson()
+    {
+        var result = _hasChangedSinceLastTick
+            ? StableJson.Changed(_serializer.Serialize().Content)
+            : StableJson.NoChanges();
+        return result;
     }
 }
