+++ GearBox.Core/Model/GameObjects/ChangeTracking/IDynamic.cs
@@ -6,4 +6,9 @@ public interface IDynamic
     /// Values on this game object which can change over time.
     /// </summary>
     IEnumerable<object?> DynamicValues { get; }
+
+    /// <summary>
+    /// Serializes this object when it changes so it can be sent to the client
+    /// </summary>
+    string Serialize(SerializationOptions options);
 }
