+++ GearBox.Core/Model/Items/Inventory.cs
@@ -11,15 +11,12 @@ namespace GearBox.Core.Model.Items;
 public class Inventory : IDynamic
 {
     private readonly ChangeTracker _changeTracker;
-    private bool _updatedLastFrame = true;
 
     public Inventory()
     {
-        Serializer = new("inventory", Serialize);
         _changeTracker = new(this);
     }
 
-    public Serializer Serializer { get; init; }
     public InventoryTab<Weapon> Weapons { get; init; } = new();
     public InventoryTab<Armor> Armors { get; init; } = new();
     public InventoryTab<Material> Materials { get; init; } = new();
@@ -83,23 +80,12 @@ public class Inventory : IDynamic
         return result;
     }
 
-    public void Update()
-    {
-        _updatedLastFrame = _changeTracker.HasChanged;
-        _changeTracker.Update();
-    }
-
-    private string Serialize(SerializationOptions options)
+    public string Serialize(SerializationOptions options)
     {
         var json = new InventoryJson(Weapons.ToJson(), Armors.ToJson(), Materials.ToJson());
         return JsonSerializer.Serialize(json, options.JsonSerializerOptions);
     }
 
-    public StableJson ToJson()
-    {
-        var result = _updatedLastFrame
-            ? StableJson.Changed(Serializer.Serialize().Content)
-            : StableJson.NoChanges();
-        return result;
-    }
+    public void Update() => _changeTracker.Update();
+    public StableJson ToJson() => _changeTracker.ToJson();
 }
